refactor: 검색 로직 service폴더로 세분화, 핸들러 실행 주석 삭제
